The program now saves the population (of the real model) when Constants.Save is true.
I fixed the bit of R  code that calculated the population because it did not calculate the population (i.e. the number of alive agents) but the number of dead agents. (Plus, it only calculated them in an interval of 5000 time steps rather than through the whole run, which wasn't convenient in my opinion.)